Relocate Docs files from 'app/docs' to 'app/md'
The Docs files including layout.tsx, page.tsx, and loading.tsx have been moved from the 'app/docs' directory to the 'app/md' directory to enhance project organization. The component implementation remains the same, it's merely a relocation of the files to a more appropriate directory.
